@@ -7,8 +7,8 @@ namespace RevXPortal.API
 	public interface IProviderEndpoint
 	{
 		Task AddProvider(ProviderModel model);
-		Task<List<ProviderModel>> GetEnabled(string userId);
+		Task<List<ProviderModel>> GetEnabled();
 		Task EditProvider(ProviderModel model);
-		Task<List<ProviderModel>> GetAll(string userId);
+		Task<List<ProviderModel>> GetAll();
 	}
 }
