@@ -8,7 +8,7 @@
 <style>
     .revx-sidebar {
         background-image: linear-gradient(to bottom right, #334c66, #3e6581) !important;
-        width: 250px;
+        width: 14rem;
         height: calc(100vh - 3.5rem);
         position: sticky;
         top: 0;
@@ -284,6 +284,7 @@
                 authenticationErrorText = "Looks like the API is offline.";
                 showAuthenticationError = true;
             }
+            NavManager.NavigateTo(NavManager.Uri, forceLoad: true);
             btnText = "Sign In";
         }
     }
