@@ -6,13 +6,18 @@
             </NavLink>
         </li>
         <li class="nav-item px-3">
-            <NavLink class="nav-link" href="management">
-                <i class="fas fa-sliders-h" style="font-size: 1.4rem; margin-right: 15px"></i> Management
+            <NavLink class="nav-link" href="invoices">
+                <i class="fas fa-file-invoice-dollar" style="font-size: 1.4rem; margin-right: 15px"></i> Invoices
             </NavLink>
         </li>
         <li class="nav-item px-3">
             <NavLink class="nav-link" href="reports">
                 <i class="fas fa-chart-pie" style="font-size: 1.4rem; margin-right: 15px"></i> Reports
+            </NavLink>
+        </li>
+        <li class="nav-item px-3">
+            <NavLink class="nav-link" href="management">
+                <i class="fas fa-sliders-h" style="font-size: 1.4rem; margin-right: 15px"></i> Management
             </NavLink>
         </li>
     </ul>
